Reworked folders normalisation
Skill --> Skills
Fixed skills in "scripts" folder
