+++ Use cases/3rd Saudi Forum for 4IR 2025/Skills/remergency_thanks.py
